Fix compilation in C++ mode
After this change the code is both valid C99 and valid C++; this helps
to consume this library from C++ since you no longer have to have a .c
file to compile it - can just include it into an existing .cpp file.

Additionally this probably makes the library useful in earlier VS versions.