Implement cgltf_parse_file
This function opens the file, reads the contents into memory and parses
the result. The file data is currently preserved and is deallocated with
cgltf_free - we could sometimes deallocate it earlier but if we want to
expose access to extension data via json string ranges, we'd need this
long term anyway.