Implement support for scenes and nodes
We mirror the GLTF spec with the data structures for scenes & nodes:

- There's one main scene (it's called "scene" in GLTF spec but that name
seems ambiguous wrt "scenes"?)
- Each scene has a list of root nodes
- Each node has a list of children

While we could linearize the children references in one large array it
doesn't seem worthwhile to do so.

Each node has either a TRS set (with rotation represented by a
quaternion), or a matrix; while we could create all representations it
seems better to store them in a raw form. We use has_matrix to
disambiguate the transform representation.

Error handling during parsing doesn't work perfectly in the new code,
but it mostly mirrors the existing code in terms of checks. There will
need to be certain changes to make sure parser can cleanly exit at any
point with an error; that's outside of the scope of this change.