Implement support for skins
This change adds support for parsing skins, along with joints and
inverse bind matrix accessor, as well as skin references inside nodes.

Skin has a skeleton that is optional in GLTF spec, and in absence of a
skin is supposed to refer to "scene root" - we leave this pointer as
NULL in this case.