Add has_translation/rotation/scale
Instead of storing data in a union with one boolean has_matrix, it seems
cleaner to store all data with individual fields for
has_translation/etc.

Separate boolean fields allow users to easily query presence of TRS data
without comparing the values with identity.

GLTF spec prohibits matrix to co-exist with TRS so our new data
definition is slightly redundant but makes more sense and is easier to
extend in the future - if GLTF adds "skew" we'll just add one more
bool field and an array.