Change cgltf_parse and cgltf_parse_file to return a pointer to data
This change switches to a model where cgltf_parse variants allocate
cgltf_data instead of filling user-provided struct. This clarifies the
interface - cgltf_parse will return a valid data object only when it
returns success status, in which case cgltf_free must be used to
deallocate memory.

If a parse error happens in the middle of parsing the document, we're
going to call cgltf_free and return a failure status. This mechanism
isn't 100% robust since some pointers aren't always initialized - this
will be fixed separately.

The code was slightly restructured to delay the creation of cgltf_data
object up until many errors have already been detected, and also this
change has a few memory leak fixes.

The outer loop was also changed to the more usual "iterate through all
keys" style that is used elsewhere in the parser.