Rework pointer fixup and allocation strategy to be more robust
To make sure we can free the parse result at any point in time, we
zero-initialize all data to make sure all pointers are NULL; this also
simplifies our manual initialization code.

This introduces a conflict with index encoding scheme, so we now encode
index 0 as pointer value 1, to make sure NULL value is preserved.

To make this kind of transition easier to do in the future we now use
CGLTF_PTRINDEX and CGLTF_PTRFIXUP macros to assign indices to pointers
and fix them up. The fixup macro additionally verifies that the
generated pointer is in bounds.