Implement support for KHR_materials_pbrSpecularGlossiness
We now can parse materials that use specular/glossiness workflow. GLTF
extension spec says that information about metallic-roughness and
specular-glossiness can coexist; while this is unlikely to happen in
practice, this is why instead of an enum that picks the material
representation we have two booleans.

Also add tokens for fuzzing and update README.