Implement support for morph targets
We now parse morph targets for each mesh, morph weights for meshes,
morph weights for nodes and animation target "weights".

Morph targets use the same attribute structure as the primitive itself;
this results in a pretty deep object hierarchy :( This may be simplified
to a flat C array with accessor pointers later?