Implement support for asset information parsing
We now parse "asset" object which contains the generator (that can be
useful to figure out what pecularities of the data we should expect),
copyright and version information.

cgltf_data::version was misleading - it referred to GLB version only and
had nothing to do with GLTF version, so it was removed. However we now
check that the version matches GLB spec.

Also rename Gltf constants to Glb - again, they define the container
format, not the GLTF data encoded inside it.