Parse accessor min/max limits
Since accessors can refer to values up to 16 scalars, the maximum size
of min/max arrays is 16.

We do not always know the size of the arrays because "type" is in the
same object and may happen after the arrays, so we just parse the entire
array - up to 16 elements to not cause heap overflow.