Substantially improve parser safety for malformed inputs
This time we can run millions of iterations of llvm-fuzz without
crashing. There still are some cases where we may not parse the JSON
correctly due to lack of validation (e.g. when we expect an integer and
get a string instead, we will convert the integer to -1 and parse will
not fail). Also we're slowly leaking memory during the fuzzing process,
that's going to be fixed separately.

The key part of this change is checking keys for objects. When JSON
doesn't have correct structure for object contents, you can look at the
key and verify that its size is 1 - if it's not (it's usually 0), that's
an indication that there was no value that was paired with the key.

Because of this we introduce a macro to check the type and size of all
keys, and rigorously apply it to all object parsing loops.

Additionally, this change cleans up error propagation and fixes several
cases where it was missed. Now we check (i < 0) in the loops outside of
the if() chain.