Implement KHR_texture_transform extension
This adds straightforward parsing of the texture transform (that has to
be done for each texture view) via the extension. has_transform can be
used to check if the transform is present.

This change also moves scale initialization into parse_texture_view. If
the texture isn't present, it shouldn't matter what the scale is.