Finish extension support
Update README to list all supported extensions and update fuzz
dictionary. With all other PRs we now support all KHR extensions except
for Draco that will not be supported since it's a very large C++
dependency.

It probably doesn't make a lot of sense to support non-KHR extensions
for now.