Rework string parsing to avoid memory leaks on malformed documents
When a string element occurs twice in the input document, we allocate a
new string array over the old one, which means the old array never gets
freed.

To solve this, we now use parse_json_string function which returns a
JSON error if the key already exists; this gives us an opportunity to
easily handle the out-of-memory error as well.