Rework array parsing to avoid memory leaks on malformed documents
Similarly to previous commit, if an array occurs twice in the document,
we used to overwrite the data for the previous update, leading to memory
leaks.

We now use cgltf_parse_json_array function that checks if the array was
already allocated, and also checks for memory allocation failure.

With this, we shouldn't have any leaks on malformed documents.