Add helper functions to extract node transforms
These are helpful for applications that don't care about the specific
TRS composition of node transforms and just want the raw matrix data.

While we could go the other way as well and provide TRS, that requires
decomposing the matrix which is not as numerically stable, and would
introduce interfaces that could become invalid if GLTF adds skew as an
extension.