Implement generic support for parsing attribute index along with type
With this change instead of hardcoding the index of the attribute into
the type enum, we separate it out into an integer. This means that we
now support an arbitrary number of UV sets; additionally if any exporter
exports more than 4 bone indices using JOINTS_1/WEIGHTS_1 then that
should work as well.