Fix GLB loading when first buffer has a URI
I believe GLTF spec optionally allows the first buffer to reference the
GLB-stored buffer but it doesn't seem to require it (it makes sense to
always have this mapping since otherwise GLB buffer is orphaned, but the
spec doesn't say so). Because of this optionality we should check that
URI isn't defined for the first buffer before binding it to bin.