Fix GLB buffer handling when GLB buffer is too short
When first buffer is loaded from the GLB binary chunk, if the chunk is
too small we'd be returning a data pointer to the user while supplying a
buffer that's too short to access safely, so this has to be an error.

Additionally fix handling of buffers with no URI - URI is optional as
per spec, and presumably buffers without a URI may have their contents
supplied via extensions. It seems better to keep the buffers like that
as is, that way the application can load these manually separately.