Make parser memory-safe
This change fixes the remaining memory safety issues, at least according
to the fuzzer.

There's two changes here that are important:

- parse_json_texture_view had sloppy error handling; this change fixes
this and also cleans up pbr_specular_glossiness checks

- Even in strict mode, jsmn can generate non-sensical JSON structures
which lead to off-by-one access to the token buffer. To fix this, we
append UNDEFINED token to the token buffer which makes sure that we read
(and reject) the UNDEFINED token during subsequent parsing pass.