Fix a few corner cases in cgltf_component_read_*
This change fixes a few issues with the existing implementation:

- Normalized 8-bit unsigned integers were decoded using CHAR_MAX as the
divider; signedness of char is not specified by the standard and it's
*very* common to have char be signed (it's usually unsigned on ARM
systems and signed everywhere else), so this used 127 instead of 255

- In general it's incorrect to rely on USHRT/UCHAR/etc. limits - we are
dealing with fixed-width quantities, and in theory the short/char/etc.
could have different sizes or limits. It's more correct to hardcode the
numbers from the spec.

- glTF spec doesn't define normalization behavior for 32-bit integers.
Note that it also doesn't define a signed 32-bit integer type and it's
unlikely that one will happen without the other. This is relevant
because 2^32-1 isn't representable exactly as a float; removing the case
entirely alleviates precision concerns.

- Finally, unknown types should not be treated the same way as r8u, and
instead should return a placeholder value.