Add a separate error code for legacy (1.0) glTF files
When parsing glTF 1.0 files, it helps to differentiate the error code so
that the user can be presented with an error message urging them to
convert the file.

GLB 1.0 files are easy to detect - they have the version number 1.

glTF 1.0 files are harder to detect. Technically they have asset.version
attribute equal to 1.0, but we parse the file in order and usually the
JSON blob has other collections, such as "accessors", before "asset".
What happens then is that we try to parse the array of accessors et al,
but in glTF 1.0 collections were objects (with key associated with
object name).

We could pre-scan the entire file looking for "asset", but this is
linear in token count, and as such moderately expensive. This change
conservatively assumes that if we are looking for an array and get the
object instead, this is a legacy file.