Improve validation to catch more cases of structurally unsound files
This change adds several more validation criteria:

- We were not successfully detecting node cycles before - a node could
not have more than one parent, but a cycle was possible. This creates
issues for code that traverses children or parents recursively. We now
use the standard cycle detection method on parent chain.

- We were not checking that scene roots don't have a parent. This is
usually more or less benign but it's the only other parent/child
validity check that could make sense...

- Previously, animation accessors were not validated wrt their element
count. This is somewhat tricky especially for morph targets - it's not
really feasible to parse animation data without knowing the morph target
count! - so we now carefully compute the expected number of values in
the output stream and check that we have that many.