Add support for KHR_draco_mesh_compression
This change adds support to read and write glTF files that use
the KHR_draco_mesh_compression extension. The extension can be
specified per primitive, which is why cgltf_primitive now has 2
new fields: has_draco_mesh_compression and draco_mesh_compression.
draco_mesh_compression contains a buffer view and a list of
attributes.

It is important to note that the extension may or may not be
required. A primitive can specify both compressed and uncompressed
data. While this doesn't affect the parsing, it does affect the
writing since the proper extensionsUsed/Required needs to be
specified. To achieve this, this change adds support for required
extensions in cgltf_write (see required_extensions_flag in
cgltf_context). The extension is deemed required when a primitive
has no attributes or no indices.

Extension spec:
https://github.com/KhronosGroup/glTF/tree/master/extensions/2.0/Khronos/KHR_draco_mesh_compression