Implement cgltf_decode_uri
Image and buffer paths are encoded as URIs in the glTF file. This means
that special characters, such as spaces, are percent-encoded.

This change adds a helper function, cgltf_decode_uri, that can decode a
URI in place; it also uses it when loading the buffers so that we pass a
path to file read callback instead of a path-URI combo.