Implement getter functions for indices of various objects
This can be useful to recover the index from the pointer for cases when
another cgltf data structure contains the pointer.

For animation samplers/channels, the channel references the sampler, but
nothing references the channel; it felt inconsistent to only add the
sampler function, even though the channel function is unlikely to be
useful.