Unpack indices of arbitrary component types
Since the `memcpy` only needs the size of the index data, it can
use the component size (when equal to the stride) to determine
the total number of bytes to copy. This allows it to work for
different component types, including the common
`cgltf_component_type_r_16u`.

This also adds a reference comment for the function at the top of
the file, consistent with `cgltf_accessor_unpack_floats`.