Change error code on sampler count mismatch to invalid_gltf
This error as it stands is a little misleading: it's the only case where
we emit an error like this when the sizes aren't equal, so it can hit
when the output count is *larger* (so there's technically enough data,
but the glTF file is still invalid per spec / per Khronos validator). In
other cases where a size mismatches like this we currently emit
invalid_gltf so this seems to make more sense.