Add cgltf_find_accessor helper
When using cgltf something that comes up often is the need to find the
accessor corresponding to a given attribute for a primitive (e.g.
positions or second UV channel). The code for this is simple, but it is
needed so frequently that it would be nice to include a helper for this
in cgltf proper.