Add file size to cgltf_file_options::release callback
This is necessary for POSIX systems to be able to implement file loading
via mmap/munmap since munmap takes the size (unlike Windows). Without
this, the implementation would need to keep a map from file pointer to
file size to be able to unmap the file.

This change requires minor adjustment for people who use file callbacks
but the change should be very straightforward and will be flagged during
the upgrade by the compiler.