Implement parsing support for KHR_meshopt_compression draft
Because the JSON structure matches that of EXT_meshopt_compression,
and both structures can't be present on the same buffer view, we reuse
the existing C structure and merely tag the object "is_khr", which
could be useful to decoding -- although the updated vertex data will
be automatically decompressed with a meshoptimized update without code
changes.

In addition to vertex encoding v1, which will be decoded automatically,
KHR_meshopt_compression adds support for color filters which requires an
extra enum entry.